requirements.txt
asgiref==3.11.0
asttokens==3.0.1
black==23.11.0
boto3==1.34.7
botocore==1.34.162
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==2.1.0
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-heroku==0.3.1
django-redis==5.3.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
executing==2.2.1
factory-boy==3.3.0
Faker==23.2.0
filelock==3.20.3
flake8==6.1.0
gunicorn==21.2.0
identify==2.6.16
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipython==8.17.2
isort==5.12.0
jedi==0.19.2
jmespath==1.1.0
Markdown==3.5.2
matplotlib-inline==0.2.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.10.0
oauthlib==3.3.1
packaging==26.0
parso==0.8.5
pathspec==1.0.3
Pillow==10.1.0
platformdirs==4.5.1
pluggy==1.6.0
pre-commit==3.5.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.9
pure_eval==0.2.3
pycodestyle==2.11.1
pycparser==3.0
pycryptodome==3.20.0
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.4
setuptools==80.10.2
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.36.1
wcwidth==0.3.1
whitenoise==6.6.0
